web/scripts/FlipbookFactory/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# For future steps (install as needed)
# opencv-python>=4.8.0
# mediapipe>=0.10.0
# torch>=2.0.0
# diffusers>=0.21.0
